% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onehot.R
\name{onehot}
\alias{onehot}
\title{creates a dataframe of one hot encoding for a factor variable}
\usage{
onehot(facvar)
}
\arguments{
\item{facvar}{a categorical vector}
}
\value{
a dataframe
}
\description{
creates a dataframe of one hot encoding for a factor variable
}
\examples{
mock_data <- data.frame(fac = c(rep("A",10),rep("B",10),rep("C",10)))
onehot(mock_data$fac)
}
