% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PropRollup.R
\name{PropRollup}
\alias{PropRollup}
\title{creates a dataframe of proportions of each factor level within a group}
\usage{
PropRollup(facvar, groupvar, weightvar = rep(1, length(facvar)))
}
\arguments{
\item{facvar}{a categorical vector}

\item{groupvar}{the variable to be grouped by}

\item{weightvar}{the variable to weight by if we desire a weighted mean for the calculation}
}
\value{
a dataframe
}
\description{
creates a dataframe of proportions of each factor level within a group
}
\examples{
mock_data <- data.frame(fac = c(rep("A",10),rep("B",10),rep("C",10)),grp=c(rep("G1",15),rep("G2",15)))
PropRollup(mock_data$fac,mock_data$grp)
}
